Add test_EditConnection_Save
