+++ ios/LoggingClient/UITests/BUILD.bazel
@@ -7,6 +7,7 @@ swift_library(
     srcs = [
         "LoggingClientUITests.swift",
         "LoggingClientUITestsLaunchTests.swift",
+        "XCUIElement+Extensions.swift",
     ],
     module_name = "UITests",
     tags = ["manual"],
